Move tests to classes
