+++ .gitignore
@@ -2,6 +2,7 @@ target/
 !.mvn/wrapper/maven-wrapper.jar
 !**/src/main/**/target/
 !**/src/test/**/target/
+src/main/java/basic/experiments/
 
 ### IntelliJ IDEA ###
 .idea/uiDesigner.xml
